// src/app/enums/font-weights.ts
export enum FontWeights {
  normal = 'normal',
  bold = 'bold',
  lighter = 'lighter',
  bolder = 'bolder',
  hundred = '100',
  twoHundred = '200',
  threeHundred = '300',
  fourHundred = '400',
  fiveHundred = '500',
  sixHundred = '600',
  sevenHundred = '700',
  eightHundred = '800',
  nineHundred = '900',
  inherit = 'inherit',
  initial = 'initial',
  unset = 'unset',
}